aad-tool fails if the user has FIDO2 enabled
Fixes: https://github.com/himmelblau-idm/himmelblau/issues/393
