+++ src/common/Cargo.toml
@@ -49,3 +49,7 @@ csv = { workspace = true }
 sketching = { workspace = true }
 kanidm_lib_file_permissions.workspace = true
 regex = "1.11.1"
+sha2 = "0.10.9"
+base64.workspace = true
+authenticator = { version = "0.4.1", default-features = false, features = ["crypto_openssl"] }
+rpassword = "7.4.0"
