+++ src/pam/src/pam/constants.rs
@@ -82,42 +82,4 @@ pub const PAM_TEXT_INFO: PamMessageStyle = 4;
 pub const _PAM_RADIO_TYPE: PamMessageStyle = 5;
 pub const _PAM_BINARY_PROMPT: PamMessageStyle = 7;
 
-// The Linux-PAM return values
-// see /usr/include/security/_pam_types.h
-#[allow(non_camel_case_types, dead_code)]
-#[derive(Debug, PartialEq)]
-#[repr(C)]
-pub enum PamResultCode {
-    PAM_SUCCESS = 0,
-    PAM_OPEN_ERR = 1,
-    PAM_SYMBOL_ERR = 2,
-    PAM_SERVICE_ERR = 3,
-    PAM_SYSTEM_ERR = 4,
-    PAM_BUF_ERR = 5,
-    PAM_PERM_DENIED = 6,
-    PAM_AUTH_ERR = 7,
-    PAM_CRED_INSUFFICIENT = 8,
-    PAM_AUTHINFO_UNAVAIL = 9,
-    PAM_USER_UNKNOWN = 10,
-    PAM_MAXTRIES = 11,
-    PAM_NEW_AUTHTOK_REQD = 12,
-    PAM_ACCT_EXPIRED = 13,
-    PAM_SESSION_ERR = 14,
-    PAM_CRED_UNAVAIL = 15,
-    PAM_CRED_EXPIRED = 16,
-    PAM_CRED_ERR = 17,
-    PAM_NO_MODULE_DATA = 18,
-    PAM_CONV_ERR = 19,
-    PAM_AUTHTOK_ERR = 20,
-    PAM_AUTHTOK_RECOVERY_ERR = 21,
-    PAM_AUTHTOK_LOCK_BUSY = 22,
-    PAM_AUTHTOK_DISABLE_AGING = 23,
-    PAM_TRY_AGAIN = 24,
-    PAM_IGNORE = 25,
-    PAM_ABORT = 26,
-    PAM_AUTHTOK_EXPIRED = 27,
-    PAM_MODULE_UNKNOWN = 28,
-    PAM_BAD_ITEM = 29,
-    PAM_CONV_AGAIN = 30,
-    PAM_INCOMPLETE = 31,
-}
+pub use himmelblau_unix_common::pam::PamResultCode;
